+++ src/main/java/pentastagiu/repository/NotificationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pentastagiu.model.Notification;
 
 @Repository
-public interface NotificationRepository extends CrudRepository<Notification,Integer> {
+public interface NotificationRepository extends CrudRepository<Notification,Long> {
 }
